+++ pytorch_lightning/accelerators/deprecated_api.py
@@ -12,8 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pytorch_lightning.accelerators.types import DistributedType, DeviceType
-from pytorch_lightning.utilities import rank_zero_warn
+from pytorch_lightning.utilities import rank_zero_warn, DistributedType, DeviceType
 
 
 class DeprecatedDistDeviceAttributes:
